backend/pyproject.toml
[tool.poetry]
name = "inv-cook"
version = "0.1.0"
description = ""
authors = ["Sebastian Zhao <sebbyzhao@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.25.0"
scipy = "^1.11.0"
matplotlib = "^3.7.1"
torch = "^2.0.1"
torchvision = "^0.15.2"
nltk = "^3.8.1"
pillow = "^9.5.0"
tqdm = "^4.65.0"
lmdb = "^1.4.1"
tensorflow = "^2.12.0"
tensorboardx = "^2.6.1"
diffusers = "^0.17.1"
uvicorn = "^0.22.0"
fastapi = "^0.98.0"
pydantic = "^1.10.9"
python-multipart = "^0.0.6"
transformers = "^4.30.2"


[tool.poetry.group.dev.dependencies]
pip = "^23.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
